requirements.txt
anthropic
bcrypt
build
discord.py
diskcache
email-validator
fastapi
gunicorn
jinja2
langdetect
litellm>=1.59.6,<1.59.9
loguru
openai
pandas
psycopg2-binary
pydantic_settings
pyjwt
pytest-asyncio
pytest-mock
pytest-xdist
rich
sendgrid
sentry-sdk[fastapi]
setuptools
tenacity
tiktoken
typer
uvicorn
wheel
zxcvbn